removing some moved components
