+++ src/components/BookList.js
@@ -1,6 +1,34 @@
 import React, {useState, useEffect} from 'react';
-import getBookList from "../utils/getBookList";
-import BookLink from "./BookLink";
+import {Link} from "react-router-dom";
+import {FaCheckCircle} from 'react-icons/fa';
+import { MdDoNotDisturb } from "react-icons/md";
+import documentClient from '../configuredDocumentClient';
+
+async function getBookList() {
+    return await documentClient.scan({
+        TableName: process.env.REACT_APP_BOOK_TABLE
+    }).promise();
+}
+
+const isAvailable = (isHere) => isHere ?
+        (<label><FaCheckCircle color="#2AF598" size={16} /> Available </label>) :
+        (<label><MdDoNotDisturb color="#FF0000" size={16} /> Not Available </label>);
+
+const BookLink = (props) => {
+    const {bookId, title, shelf, isAvailable: available } = props.book;
+    return (
+        <div>
+            <div>
+                <Link to={`/books/${bookId}`}>{title}</Link>
+            </div>
+            <p>
+                {isAvailable(available)}
+                on Shelf {shelf}
+                <hr/>
+            </p>
+        </div>
+    );
+}
 
 const BookList = () => {
     const [loading, setLoading] = useState(true);
